refactor(infrastructure): split FeaturesLayer into focused services
Extract FeaturesLayer monolith (654 lines) into 5 specialized services
following the same orchestrator pattern as other layers.

New services in packages/infrastructure/src/map/services/layers/features/:
- HazardPlacementService: placeHazards() — terrain-condition-based hazard
  placement (quicksand, unstable ground, poison plants, animal dens)
- ResourcePlacementService: placeResources() — environment-based resource
  placement (medicinal herbs, berries, fresh water, mineral deposits)
- LandmarkPlacementService: placeLandmarks() + findRidges() — narrative POI
  placement (ancient trees, standing stones, cave entrances, battlefield remains)
- TacticalFeaturePlacementService: identifyTacticalFeatures() + findHighPoints()
  + findChokePoints() — combat-relevant feature identification (high ground,
  choke points, ambush sites, vantage points)
- FeatureTileGenerationService: createTiles() + hazard/resource/tactical
  descriptions — priority-layered tile assembly with description generation

FeaturesLayer becomes a lightweight orchestrator (~110 lines):
- Injects 5 services via class-based @Inject() tokens (no string tokens)
- @Optional() ILogger for structured logging
- Width/height derived from geology layer; passed as params to each service
- Unused context param renamed to _context (OxLint compliance)

Register 5 new providers in infrastructure.module.ts under
'Features internal services' section.
