+++ packages/infrastructure/src/map/services/layers/features/index.ts
@@ -1,0 +1,5 @@
+export { HazardPlacementService } from './HazardPlacementService';
+export { ResourcePlacementService } from './ResourcePlacementService';
+export { LandmarkPlacementService } from './LandmarkPlacementService';
+export { TacticalFeaturePlacementService } from './TacticalFeaturePlacementService';
+export { FeatureTileGenerationService } from './FeatureTileGenerationService';
