+++ packages/infrastructure/src/map/services/layers/index.ts
@@ -31,3 +31,5 @@ export * from './structures';
 
 // Features and points of interest layer implementation
 export { FeaturesLayer } from './FeaturesLayer';
+// Features internal services
+export * from './features';
